New feature `use asdf current` to activate current plugins. (#14)
* New feature `use asdf current` to activate current plugins.

In your `.envrc` file now you can do: `use asdf current`.
And it will load the same plugins and versions as listed by
the output of `asdf current`.

Notice that `use asdf current` is a bit slow since it will
try to determine the version for each installed plugin
(same thing that `asdf current` does). But, the upside is
that you will have not only local plugins active but also
those in global ~/.tool-versions file (as resolved by asdf).

This means `use asdf current` will also activate plugins
that use legacy files and thus are not present inside
`.tool-versions`. (this was a limitation of `use asdf local`)

The old behaviour is now named `use asdf local` and is
currently still the default if you give no args to `use asdf`.
This is because `use asdf current` is slower, and I'm not
sure all people would like that by default.

* Add some failing tests

* Working on basic tests

* Add use asdf tests

* Add more usage details on README

* Mention README on invalid use_asdf usage