Custom plugin shims now lower in PATH than bins
We no longer use asdf's `with_plugin_env`, since
it by default prepends custom-shims higher in path
which caused some issues as documented on #82.

We still add the shims directory (we had a test for
this) but now just after the actual installed bins.

We also use `source_env` to load the plugin's `exec-env`
with the added benefit that direnv will watch the file
for changes and re-trigger if people upgrade the plugin.

Fixes #82