Add `ASDF_DIRENV_DEBUG` to the tools file checksum.
In https://github.com/asdf-community/asdf-direnv/pull/110/, we changed
things so the generated envrc file is different based on if the
`ASDF_DIRENV_DEBUG` environment variable is set. Before this change, if
you were to generate a cached file while this env var is set, when you
switch the env var off, you'd *still* have a `set -x` in your cached
envrc file.