RFC: Add a new `asdf direnv shell` command
Last week, I learned that @chris-lasher-honor used to be an
`asdf-direnv` user, but switched back to plain `asdf` because he liked
using `asdf shell` for small one-off commands like `npx
create-react-app`. I think that's a pretty solid use-case where it's not
necessarily worth the overhead of creating a `.tool-versions` file and a
`.envrc` file just to create a brand new folder that's going to also needs
its own set of the same dotfiles.

I also sometimes find that the activation energy necessary just to play
around with a random tool is a bit high with asdf-direnv:

- create a new directory
- run `asdf local <plugin> <version>`
- create a `.envrc` file
- run `direnv allow`

Now, you can do something like `asdf shell python 3.8.10`, and be up and
running immediately:

    $ asdf current
    direnv          2.30.3          /home/jeremy/.tool-versions
    nodejs          ______          No version is set. Run "asdf <global|shell|local> nodejs <version>"
    $ node --version
    bash: node: command not found

    $ asdf direnv shell nodejs 10.17.0
    Loading nodejs 10.17.0
    direnv: loading ~/.asdf/plugins/nodejs/bin/exec-env

    $ asdf current
    direnv          2.30.3          /home/jeremy/.tool-versions
    nodejs          10.17.0         ASDF_NODEJS_VERSION environment variable
    $ node --version
    v10.17.0