Support 'latest:X' in .tool-versions
asdf supports specificying a version like 'latest:3' which means the
latest stable version, in the 3.x.y series of a program. This adds
support for this feature in this plugin.

Also adds a test, and support for 'list-all' in the test dummy plugin.
'list-all' is called by asdf internally when using the 'latest:X'
format.