disable nounset while using asdf's find_versions
since it expects ASDF_DEFAULT_TOOL_VERSIONS_FILENAME to be defined.

Probably, we should send a patch for it to asdf-core. However,
I believe all of asdf's `utils.bash` should be considered private API
and no one but asdf-core should rely on it. Yet we do because we want
to resolve versions, paths and environments exactly as asdf does.

Thanks to @devinus for reporting at https://github.com/asdf-community/asdf-direnv/pull/152#issuecomment-1107759268